moved Parser marshaling functions ictiobus -> parse
